% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_equation.R
\name{generate_equation}
\alias{generate_equation}
\title{Generate a random equation}
\usage{
generate_equation()
}
\value{
A two column row with a LaTeX formula and description.
}
\description{
Generate a random equation
}
\examples{
generate_equation()
}
